pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "tablemage"
version = "0.1.0a1"
description = "A Python package for low-code analysis of tabular data"
authors = [
    {name = "Andrew Yang", email = "andrew_j_yang@brown.edu"},
    {name = "Joshua Woo", email = "joshua_woo@brown.edu"},
    {name = "Ryan Zhang", email = "rzryanz00@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.3",
    "pandas>=2.2.2",
    "scipy>=1.1.3",
    "scikit-learn==1.5.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "xgboost>=2.0.3",
    "statsmodels>=0.14.0",
    "adjustText>=1.1.1",
    "tableone>=0.9.1",
    "umap-learn>=0.5.7",
    "optuna>=3.6.1",
    "optuna-integration>=3.6.0",
    "tableone>=0.9.1"
]
license = {file = "LICENSE"}
keywords = ["data science", "low-code", "no-code", "statistics", "machine learning"]
classifiers = [
    "Programming Language :: Python"
]


[project.optional-dependencies]
agents = [
    "llama-index>=0.11.23",
    "llama-index-experimental>=0.4.0",
    "llama-index-llms-groq>=0.2.0",
    "llama-index-llms-ollama>=0.3.6",
    "flask",
    "python-dotenv",
    "llama-index-embeddings-fastembed",
    "llama-index-llms-ollama",
    "fastembed",
]


[project.urls]
Homepage = "https://tablemage.readthedocs.io/"


[tool.hatch.metadata]
packages = [
    {include = "tablemage"}
]
